Unbreak vmx(4) on i386.  The right shift of 32 bits for the DSH register (driver shared address high) exceeded the width of the 32bit bus address; casting the address to a 64bit type will correctly result in a zero value on i386 and the high bits on amd64.

+++ sys/dev/pci/if_vmx.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: if_vmx.c,v 1.11 2013/06/22 00:28:10 uebayasi Exp $	*/
+/*	$OpenBSD: if_vmx.c,v 1.12 2013/08/28 10:19:19 reyk Exp $	*/
 
 /*
  * Copyright (c) 2013 Tsubai Masanari
@@ -366,7 +366,7 @@ vmxnet3_dma_init(struct vmxnet3_softc *sc)
 	for (i = 0; i < VMXNET3_NINTR; i++)
 		ds->modlevel[i] = UPT1_IMOD_ADAPTIVE;
 	WRITE_BAR1(sc, VMXNET3_BAR1_DSL, ds_pa);
-	WRITE_BAR1(sc, VMXNET3_BAR1_DSH, ds_pa >> 32);
+	WRITE_BAR1(sc, VMXNET3_BAR1_DSH, (u_int64_t)ds_pa >> 32);
 	return 0;
 }
